requirements.txt
django
djangorestframework
selenium
bs4